backend/requirements.txt
fastapi>=0.111,<1.0
uvicorn[standard]>=0.29,<1.0
celery[redis]>=5.3,<6.0
redis>=5.0,<6.0
pydantic-settings>=2.0,<3.0
python-jose[cryptography]>=3.3,<4.0
passlib[bcrypt]>=1.7,<2.0
httpx>=0.27,<1.0
